requirements.txt
apache-airflow==2.8.1
pandas
scikit-learn
joblib
requests
boto3
python-dotenv